+++ src/modules/car/car.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common'
 import { PrismaService } from 'src/services/prisma.service'
 
-import { CarService } from '../../services/car.service'
 import { CarResolver } from './car.resolver'
+import { CarService } from './car.service'
 
 @Module({
   providers: [PrismaService, CarResolver, CarService]
