+++ src/modules/car/dto/index.ts
@@ -1,0 +1,3 @@
+export * from './create_car.input'
+export * from './create_manufacturer.input'
+export * from './create_photo.input'
